+++ pokedex/src/App.js
@@ -1,5 +1,6 @@
 import React from "react";
-import Router from "./Router/Router"
+import Router from "./Router/Router";
+
 
 const App = () => {
   return <Router />;
